Correction in class MinimumSecuritySettingsSupplierFactory to pass checkstyle tests, licence header added where missing, test classes moved to proper packages.
