+++ src/integrationTest/java/org/opensearch/security/SecurityRolesTests.java
@@ -24,7 +24,7 @@
  * GitHub history for details.
  */
 
-package org.opensearch.test;
+package org.opensearch.security;
 
 import com.carrotsearch.randomizedtesting.annotations.ThreadLeakScope;
 import org.apache.http.HttpStatus;
